+++ subs/client.go
@@ -95,7 +95,10 @@ func (h *SocketEventHandler) handleConfigSchema(conn etp.Conn, data []byte) []by
 	h.store.VisitReadonlyState(func(readState state.ReadonlyState) {
 		configs = readState.Configs().GetByModuleIds([]string{module.Id})
 	})
-	if len(configs) == 0 && len(configSchema.DefaultConfig) > 0 {
+	if len(configs) == 0 { //create empty default config
+		if configSchema.DefaultConfig == nil {
+			configSchema.DefaultConfig = make(map[string]interface{})
+		}
 		config := entity.Config{
 			Id:        state.GenerateId(),
 			Name:      module.Name,